+++ src/index.ts
@@ -1,12 +1,3 @@
-import { content } from './content';
-import { construction } from './construction';
-import { editor } from './editor';
-import { elections } from './election';
-import { finance } from './finance';
-import { health } from './health';
-import { logos } from './logos';
-import { programming } from './programming';
-import { socialMedia } from './socialMedia';
 import type { IMXIcon } from './common/icon';
 
 export * from './content';
@@ -19,17 +10,7 @@ export * from './logos';
 export * from './programming';
 export * from './socialMedia';
 
+export { all } from './all';
+
 export { IconCategory, IconCategoryEnum } from './common/icon';
 export type { IMXIcon };
-
-export const all: IMXIcon[] = [
-  ...content,
-  ...construction,
-  ...editor,
-  ...elections,
-  ...finance,
-  ...health,
-  ...logos,
-  ...programming,
-  ...socialMedia,
-];
